requirements.txt
PyYAML
python-telegram-bot
caldav
pillow
ticktick-py